pyproject.toml
[project]
name = "ens-retro-data"
version = "0.1.0"
description = "Data infrastructure for ENS DAO retrospective evaluation"
requires-python = ">=3.11"
dependencies = [
    "dagster>=1.9",
    "dagster-webserver>=1.9",
    "duckdb>=1.1",
    "pandas>=2.2",
    "polars>=1.0",
    "pyarrow>=18.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["infra"]

[tool.dagster]
module_name = "infra.definitions"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
